Add authorization system

+++ v1/controller/task.php
@@ -21,8 +21,90 @@ try{
     exit;
 }
 
+// begin auth script
+
+if(!isset( $_SERVER['HTTP_AUTHORIZATION'] ) || strlen( $_SERVER['HTTP_AUTHORIZATION'] ) < 1 ){
+    $response = new Response();
+
+    $response->setHttpStatusCode(401);
+    $response->setSuccess(false);
+    $response->setMessage("access token is missing from the header");
+    $response->send();
+    exit;
+}
+
+$accesstoken = $_SERVER['HTTP_AUTHORIZATION'];
+
+try{
+    $query = $writeDB->prepare("SELECT userid, accesstokenexpiry, useractive, loginattempts FROM tblsessions, tblusers
+                                WHERE tblsessions.userid = tblusers.id
+                                AND accesstoken = :accesstoken");
+    $query->bindParam(":accesstoken", $accesstoken, PDO::PARAM_STR);
+    $query->execute();
+
+    $rowCount = $query->rowCount();
+
+    if($rowCount === 0){
+        $response = new Response();
+
+        $response->setHttpStatusCode(401);
+        $response->setSuccess(false);
+        $response->setMessage("Access Token is not valid");
+        $response->send();
+        exit;
+    }
+
+    $row = $query->fetch(PDO::FETCH_ASSOC);
+
+    $returned_userid = $row['userid'];
+    $returned_accesstokenexpiry = $row['accesstokenexpiry']; 
+    $returned_useractive = $row['useractive'];
+    $returned_loginattempts = $row['loginattempts'];
+
+    if($returned_useractive !== "Y"){
+        $response = new Response();
+
+        $response->setHttpStatusCode(401);
+        $response->setSuccess(false);
+        $response->setMessage("User Account is not active");
+        $response->send();
+        exit;
+    }
+
+    if($returned_loginattempts >= 3){
+        $response = new Response();
+
+        $response->setHttpStatusCode(401);
+        $response->setSuccess(false);
+        $response->setMessage("User Account is locked out");
+        $response->send();
+        exit;
+    }
+
+    if(strtotime($returned_accesstokenexpiry) < time()){ // current time is bigger than access token expiry time
+        $response = new Response();
+            
+        $response->setHttpStatusCode(401);
+        $response->setSuccess(false);
+        $response->setMessage("Access Token has expiry - Please Log in again");
+        $response->send();
+        exit;
+    }
+}
+catch(PDOException $ex){
+    $response = new Response();
+
+    $response->setHttpStatusCode(500);
+    $response->setSuccess(false);
+    $response->setMessage("There is an Auth issue - Please try again");
+    $response->send();
+    exit;
+}
+
+// end auth script
+
 // check if taskid is exist 
-if(array_key_exists("taskid",$_GET)){
+if(array_key_exists("taskid",$_GET)){ // ex: task.php?taskid=3
 
     $taskid = $_GET['taskid'];
 
@@ -36,12 +118,13 @@ if(array_key_exists("taskid",$_GET)){
         exit;
     }
 
-    if($_SERVER['REQUEST_METHOD'] === 'GET'){
+    if($_SERVER['REQUEST_METHOD'] === 'GET'){ 
 
         try{
 
-            $query = $readDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks where id = :taskid"); // : before taskid for injection and security
+            $query = $readDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks where id = :taskid AND userid = :userid"); // : before taskid for injection and security
             $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+            $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -101,8 +184,9 @@ if(array_key_exists("taskid",$_GET)){
 
         try{
 
-            $query = $writeDB->prepare("DELETE FROM tbltasks WHERE id = :taskid");
+            $query = $writeDB->prepare("DELETE FROM tbltasks WHERE id = :taskid AND userid = :userid");
             $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+            $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -200,8 +284,9 @@ if(array_key_exists("taskid",$_GET)){
                 exit;
             }
 
-            $query = $writeDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks where id = :taskid");
+            $query = $writeDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks where id = :taskid AND userid = :userid");
             $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+            $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -220,7 +305,7 @@ if(array_key_exists("taskid",$_GET)){
                 $task = new Task($row['id'],$row['title'],$row['description'],$row['deadline'],$row['completed']);
             }
 
-            $queryString = "UPDATE tbltasks SET ".$queryField." WHERE id = :taskid";
+            $queryString = "UPDATE tbltasks SET ".$queryField." WHERE id = :taskid AND userid = :userid";
             $query = $writeDB->prepare($queryString);
 
             if($title_updated === true){
@@ -248,7 +333,7 @@ if(array_key_exists("taskid",$_GET)){
             }
 
             $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
-
+            $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -263,8 +348,9 @@ if(array_key_exists("taskid",$_GET)){
                 exit;
             }
 
-            $query = $readDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks where id = :taskid");
+            $query = $readDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks where id = :taskid AND userid = :userid");
             $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+            $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -348,8 +434,9 @@ if(array_key_exists("taskid",$_GET)){
 
         try{
 
-            $query = $readDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks where completed = :completed"); // : before taskid for injection and security
+            $query = $readDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks where completed = :completed AND userid = :userid"); // : before taskid for injection and security
             $query->bindParam(':completed', $completed, PDO::PARAM_STR);
+            $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -424,7 +511,8 @@ if(array_key_exists("taskid",$_GET)){
 
         try{
 
-            $query = $readDB->prepare("SELECT count(id) as totalNoOfTasks FROM tbltasks"); 
+            $query = $readDB->prepare("SELECT count(id) as totalNoOfTasks FROM tbltasks WHERE userid = :userid");
+            $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT); 
             $query->execute();
 
             $row = $query->fetch(PDO::FETCH_ASSOC);
@@ -448,7 +536,8 @@ if(array_key_exists("taskid",$_GET)){
 
             $offset = ($page == 1 ? 0 : ($limitPerPage*($page-1))); // offset is to know which row start with in database after first page
 
-            $query = $readDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks LIMIT :pglimit offset :offset"); // offset which row start with
+            $query = $readDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks WHERE userid = :userid LIMIT :pglimit offset :offset"); // offset which row start with
+            $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
             $query->bindParam(':pglimit', $limitPerPage, PDO::PARAM_INT);
             $query->bindParam(':offset', $offset, PDO::PARAM_INT);
             $query->execute();
@@ -511,13 +600,14 @@ if(array_key_exists("taskid",$_GET)){
         exit;
     }
 
-}elseif(empty($_GET)){ // if GET is empty mean we get all tasks not specific id or complete
+}elseif(empty($_GET)){ // if GET is empty mean we get all tasks not specific id or complete // ex: task.php
     // /tasks
     if($_SERVER['REQUEST_METHOD'] === 'GET'){
 
     try{
 
-        $query = $readDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks");
+        $query = $readDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks WHERE userid = :userid");
+        $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
         $query->execute();
 
         $rowCount = $query->rowCount();
@@ -608,11 +698,12 @@ if(array_key_exists("taskid",$_GET)){
             $completed = $newTask->getCompleted();
 
             //Insert data into database
-            $query = $writeDB->prepare("INSERT INTO tbltasks ( title, description, deadline, completed ) VALUES (:title, :description, STR_TO_DATE(:deadline, '%d/%m/%Y %H:%i'), :completed)");
+            $query = $writeDB->prepare("INSERT INTO tbltasks ( title, description, deadline, completed, userid) VALUES (:title, :description, STR_TO_DATE(:deadline, '%d/%m/%Y %H:%i'), :completed, :userid)");
             $query->bindParam(':title', $title, PDO::PARAM_STR);
             $query->bindParam(':description', $description, PDO::PARAM_STR);
             $query->bindParam(':deadline', $deadline, PDO::PARAM_STR);
             $query->bindParam(':completed', $completed, PDO::PARAM_STR);
+            $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
             $query->execute();
 
             $rowCount = $query->rowCount();
@@ -629,8 +720,9 @@ if(array_key_exists("taskid",$_GET)){
 
             $lastTaskID = $writeDB->lastInsertId();
 
-            $query = $readDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks where id = :taskid"); 
+            $query = $readDB->prepare("SELECT id, title, description, DATE_FORMAT(deadline, '%d/%m/%Y %H:%i') as deadline, completed from tbltasks where id = :taskid AND userid = :userid"); 
             $query->bindParam(':taskid', $lastTaskID, PDO::PARAM_INT);
+            $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT); 
             $query->execute();
 
             $rowCount = $query->rowCount();
